RND-767 | Simplified status sorting, separated styled components into separate files.
