+++ src/js/components/FragmentList/fragment.jsx
@@ -16,51 +16,8 @@
 
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
+import Styled from './fragment.style';
 import NodeList from './nodeList';
-
-const StyledFragment = styled.div`
-    display: flex;
-    border-bottom: 1px solid ${({ theme }) => theme.borderColor};
-    color: ${({ theme }) => theme.textColor};
-    background-color: ${({ expanded, theme }) => (expanded ? theme.fragmentHighlight : '')};
-
-    &:nth-child(2n + 1) {
-      background-color: ${({ theme }) => theme.oddFragmentBgColor};;
-      background-color: ${({ expanded, theme }) => (expanded ? theme.fragmentHighlight : '')};
-    }
-
-    &:hover, &:focus{
-      background-color: ${({ theme }) => theme.fragmentHighlight};
-    }
-`;
-
-const StatusWrapper = styled.div`
-    height: 15px;
-    width: 15px;
-    margin: 2px;
-`;
-
-const Status = styled.div`
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-
-    background-color: ${({ status }) => (status === 'success' ? '#01a101' : '')};
-    background-color: ${({ status }) => (status === 'error' ? '#ff0000' : '')};
-    background-color: ${({ status }) => (status === 'unprocessed' ? '#a7a7a7' : '')};
-    background-color: ${({ status }) => ((status === 'other' || status === 'missing') ? '#ffbb00' : '')};
-`;
-
-const Id = styled.div`
-    overflow: hidden;
-    flex: 1;
-    border-right: 1px solid ${({ theme }) => theme.borderColor};
-`;
-
-const Type = styled.div`
-    flex: 1;
-`;
 
 const Fragment = ({
   status, id, type, nodes,
@@ -79,21 +36,21 @@ const Fragment = ({
   }
 
   return (
-    <StyledFragment
+    <Styled.Fragment
       tabIndex="0"
       onClick={handleClick}
       onKeyDown={handleEnter}
       expanded={expanded}
     >
-      <StatusWrapper>
-        <Status status={status} />
-      </StatusWrapper>
-      <Id>
+      <Styled.StatusWrapper>
+        <Styled.Status status={status} />
+      </Styled.StatusWrapper>
+      <Styled.Id>
         {id}
         <NodeList expanded={expanded}>{nodes}</NodeList>
-      </Id>
-      <Type>{type}</Type>
-    </StyledFragment>
+      </Styled.Id>
+      <Styled.Type>{type}</Styled.Type>
+    </Styled.Fragment>
   );
 };
