+++ src/js/components/FragmentList/fragmentList.jsx
@@ -16,42 +16,8 @@
 
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
-import styled from 'styled-components';
+import Styled from './fragmentList.style';
 import Fragment from './fragment';
-
-const StyledFragmentList = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-`;
-
-const SortingWrapper = styled.div`
-    display: flex;
-`;
-
-const SortingButton = styled.button`
-      border: 0;
-      margin: 2px;
-      padding: 2px;
-      color: ${({ theme }) => theme.textColor};
-      background-color: ${({ theme }) => theme.buttonBgColor};
-      &:active {
-        background-color: ${({ theme }) => theme.fragmentHighlight};
-      }
-`;
-
-const SortingButtonStatus = styled(SortingButton)`
-      width: 15px;
-`;
-
-const SortingButtonId = styled(SortingButton)`
-      flex: 1;
-`;
-
-const SortingButtonType = styled(SortingButton)`
-      flex: 1;
-`;
-
 
 const FragmentList = () => {
   function mapDataToComponents({ fragments }) {
@@ -88,41 +54,12 @@ const FragmentList = () => {
   const [fragments, setFragments] = useState(sortedFragmentsByTime);
 
   function sortByStatus() {
-    function getStatusSortWeight(status) {
-      let sortWeight = 0;
-
-      switch (status) {
-        case 'error':
-          sortWeight = 0;
-          break;
-        case 'other':
-          sortWeight = 1;
-          break;
-        case 'missing':
-          sortWeight = 1;
-          break;
-        case 'success':
-          sortWeight = 2;
-          break;
-        case 'unprocessed':
-          sortWeight = 3;
-          break;
-        default: sortWeight = 4;
-          break;
-      }
-      return sortWeight;
-    }
-
-    const sortedFragments = fragments.concat().sort((a, b) => {
-      const { status: statusA } = a.props;
-      const { status: statusB } = b.props;
-      const sortWeightA = getStatusSortWeight(statusA);
-      const sortWeightB = getStatusSortWeight(statusB);
-
-      if (sortWeightA < sortWeightB) return -1;
-      if (sortWeightA > sortWeightB) return 1;
-      return 0;
-    });
+    const sortOrder = ['success', 'other', 'missing', 'unprocessed', 'error'];
+    const ordering = {};
+    sortOrder.forEach((element, index) => { ordering[element] = index; });
+    const sortedFragments = fragments.concat().sort((a, b) => (
+      ordering[a.props.status] - ordering[b.props.status] || a.props.status.localeCompare(b.props.status)
+    ));
     setFragments(sortedFragments);
   }
 
@@ -137,14 +74,14 @@ const FragmentList = () => {
   }
 
   return (
-    <StyledFragmentList>
-      <SortingWrapper>
-        <SortingButtonStatus onClick={sortByStatus}>&darr;</SortingButtonStatus>
-        <SortingButtonId onClick={sortById}>ID&darr;</SortingButtonId>
-        <SortingButtonType onClick={sortByType}>TYPE&darr;</SortingButtonType>
-      </SortingWrapper>
+    <Styled.FragmentList>
+      <Styled.SortingWrapper>
+        <Styled.SortingButtonStatus onClick={sortByStatus}>&darr;</Styled.SortingButtonStatus>
+        <Styled.SortingButtonId onClick={sortById}>ID&darr;</Styled.SortingButtonId>
+        <Styled.SortingButtonType onClick={sortByType}>TYPE&darr;</Styled.SortingButtonType>
+      </Styled.SortingWrapper>
       {fragments}
-    </StyledFragmentList>
+    </Styled.FragmentList>
   );
 };
